% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Frequency Table}
\usage{
freq_table(x, ..., cut_fmt = NULL, na_last = FALSE,
  decreasing = FALSE)
}
\arguments{
\item{x}{a \code{data.frame} or \code{data.table} object.}

\item{...}{The columns to use to create the frequency table, all combinations of the unique
levels of the columns will be used to create the rows of the frequenct table.}

\item{cut_fmt}{a function used to cut a numeric column into descrete levels
to use in the table. This parameters currently only supports a single numeric column.
Default is set to \code{NULL}. \code{cut_fmt} supports the following
predefined functions:
\itemize{
   \item Set \code{cut_fmt = "decile"} to cut the numeric column into decile breaks.
   \item Set \code{cut_fmt = "quantile"} to cut the numneric column into quantile breaks.
}}

\item{na_last}{set to \code{TRUE} to have any \code{NA} values that may be present in the
provided columns to be placed in the last row of the frequency table. Default is set
to \code{FALSE}.}

\item{decreasing}{if set to \code{TRUE} the unique levels of the columns will be presented
in a decreasing order. Default is set to \code{FALSE}.}
}
\value{
A frequency table of the unique levels of the provided columns
}
\description{
\code{freq_table} creates a simple frequency table of one or more
   columns.
}
