% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_functions.R
\name{output_to_excel}
\alias{output_to_excel}
\title{Tables to excel}
\usage{
output_to_excel(x, wb = NULL, sheet = NULL, start_row = 1,
  start_col = 1, header_style = NULL, fmt_side = TRUE,
  open_wb = TRUE, pct_keys = "percent|pct|\%|rate",
  auto_col_width = TRUE, cond_fmt_cols = NULL, keep_na = FALSE,
  rows_btwn = 2, verbose = TRUE)
}
\arguments{
\item{x}{an list of or single object of class \code{data.frame}, \code{data.table}
, \code{matrix} or \code{report.table} to write out to excel.}

\item{wb}{a workbook object created with \code{\link[openxlsx]{createWorkbook}}
using code such as \code{openxlsx::createWorkbook()}.
If no workbook object is set one will be created in the function.}

\item{sheet}{the name of a sheet created in the workbook object created using
function \code{\link[openxlsx]{addWorksheet}} using code such as
\code{openxlsx::addWorksheet(wb = wb, sheetName = <name of x>)}, where the
name of the sheet will be the name of the object passed into x}

\item{start_row}{the row to start writing the table or list of tables to.}

\item{start_col}{the column to start writing the table or list of tables to.}

\item{header_style}{a style created using \code{openxlsx::createStyle} to apply to
the header of the table}

\item{fmt_side}{if \code{header_side} should also be applied to the first columns
of the table of list of tables be output. Default is set to \code{TRUE}.}

\item{open_wb}{if the workbook object should be opened after the table has been
written.}

\item{pct_keys}{a regex argument to use to search for columns that should be
automatically formatted as percents when the table is output to excel.
Setting this function to \code{NULL} will result in no columns being
formatted as percents. Default is set to \code{"percent|pct|\%|rate"}.}

\item{auto_col_width}{automattically set column widths of the tables being
output.}

\item{cond_fmt_cols}{columns of the tables to apply conditional formatting
to.}

\item{keep_na}{if set to \code{FALSE} cells that contain \code{NA} will be
left blank in the excel document. Otherwise "NA" will be written out as
a string in excel. Default is \code{FALSE}.}

\item{rows_btwn}{if \code{x} is a list of tables the number if rows
to place between tables.}

\item{verbose}{if \code{TRUE} print the progress of the table output.
Default is set to \code{TRUE}.}
}
\description{
Write out data in table format easily to excel
}
