% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_functions.R
\name{output_to_xl.data.frame}
\alias{output_to_xl.data.frame}
\title{Output an object of class data.frame to excel}
\usage{
\method{output_to_xl}{data.frame}(x, wb = NULL, sheet = NULL,
  start_row = 1, start_col = 1, header_style = NULL,
  header_color = "#d3daea", fmt_side = TRUE, open_wb = TRUE,
  pct_keys = "percent|pct|\%|rate", auto_col_width = TRUE,
  cond_fmt_cols = NULL, keep_na = FALSE, verbose = TRUE)
}
\description{
Output an object of class data.frame to excel
}
