% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_functions.R
\name{output_to_xl.list}
\alias{output_to_xl.list}
\title{Output to excel for a list object}
\usage{
\method{output_to_xl}{list}(x, wb = NULL, sheet = NULL,
  start_row = 1, start_col = 1, header_style = NULL,
  header_color = "#d3daea", fmt_side = TRUE, open_wb = TRUE,
  pct_keys = "percent|pct|\%|rate", auto_col_width = TRUE,
  cond_fmt_cols = NULL, keep_na = FALSE, verbose = TRUE)
}
\description{
Output to excel for a list object
}
