% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.report.table}
\alias{print.report.table}
\title{Print report.table
S3 print method for objects of class report.table}
\usage{
\method{print}{report.table}(x, ...)
}
\description{
Print report.table
S3 print method for objects of class report.table
}
