% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tools.R
\docType{package}
\name{report.tools}
\alias{report.tools}
\alias{report.tools-package}
\title{report.tools: A package of tools for creating analytical reports}
\description{
The report.tools package provides functions for easily creating
analytic reports and outputting them to excel.
}
