% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar_table.R
\name{bivar_table}
\alias{bivar_table}
\title{Bivariate Table}
\usage{
bivar_table(x, var, bivar, extra_vars = NULL, cut_fmt = NULL,
  na_last = FALSE, decreasing = FALSE)
}
\arguments{
\item{x}{a \code{data.frame} or \code{data.table} object.}

\item{var}{a variable to break up the data, either all unique levels will
be used, or binned levels if a \code{cut_fmt} is chosen.}

\item{bivar}{a binary performance vector to destribute along \code{var}}

\item{extra_vars}{extra binary variables to distribute in addition to the \code{bivar}.}

\item{cut_fmt}{a function used to cut a numeric column into descrete levels
to use in the table.
Default is set to \code{NULL}. \code{cut_fmt} supports the following
predefined functions:
\itemize{
   \item Set \code{cut_fmt = "decile"} to cut the numeric column into decile breaks.
   \item Set \code{cut_fmt = "quantile"} to cut the numneric column into quantile breaks.
}}

\item{na_last}{set to \code{TRUE} to have any \code{NA} values that may be present in the
provided columns to be placed in the last row of the bivariate table. Default is set
to \code{FALSE}.}

\item{decreasing}{if set to \code{TRUE} the unique levels of the columns will be presented
in a decreasing order. Default is set to \code{FALSE}.}
}
\value{
A bivariate table with the performance field distributed along the levels of the
   selected variable.
}
\description{
Create a table that shows how a binary field distributes along another
discrete variable. Supports all unique values in a variable, custom cut
functions, or has a limited number of created helper cut functions for
continious variables.
}
\examples{
# Create a bivariate of vs by cyl
bivar_table(x = mtcars, var = "cyl", bivar = "vs")

# add extra variable am
bivar_table(x = mtcars, var = "cyl", bivar = "vs", extra_vars = "am")

# Use custom predifined cut function
bivar_table(x = mtcars, var = "mpg", bivar = "vs", extra_vars = "am", cut_fmt = "quantile")

}
