% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tools.R
\name{print.report.table}
\alias{print.report.table}
\title{S3 print method for objects of class report.table}
\usage{
\method{print}{report.table}(x)
}
\description{
S3 print method for objects of class report.table
}
